rename SearchIndexTestCase To SearchIndexCommandTestCase
